backend/requirements.txt
fastapi
uvicorn
kafka-python
pandas
scikit-learn
joblib
python-multipart
twilio==7.2.0